Add top level golang linter and golang unit test
