+++ experimental/traffic_router_golang/crconfigpoller/crconfigpoller.go
@@ -28,11 +28,11 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/cgsrch"
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/crconfig"
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/crconfigregex"
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/fetch"
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/nextcache"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/cgsrch"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/crconfig"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/crconfigregex"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/fetch"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/nextcache"
 
 	"github.com/apache/trafficcontrol/lib/go-tc"
 )
