+++ experimental/traffic_router_golang/ipmap/ipmap.go
@@ -24,7 +24,7 @@ import (
 	"errors"
 	"net"
 
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/coveragezone"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/coveragezone"
 
 	"github.com/apache/trafficcontrol/lib/go-tc"
 )
