+++ experimental/traffic_router_golang/traffic_router_golang.go
@@ -27,14 +27,14 @@ import (
 	"os"
 	"time"
 
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/availableservers"
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/config"
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/coveragezone"
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/crconfigpoller"
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/crstatespoller"
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/fetch"
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/httpsrvr"
-	"github.com/apache/trafficcontrol/traffic_router/experimental/traffic_router_golang/toutil"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/availableservers"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/config"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/coveragezone"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/crconfigpoller"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/crstatespoller"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/fetch"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/httpsrvr"
+	"github.com/apache/trafficcontrol/experimental/traffic_router_golang/toutil"
 
 	"github.com/apache/trafficcontrol/lib/go-log"
 	to "github.com/apache/trafficcontrol/traffic_ops/client"
