+++ test/router/data/httpresult.go
@@ -8,9 +8,9 @@ package data
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -18,7 +18,6 @@ package data
  * specific language governing permissions and limitations
  * under the License.
  */
-
 
 import "time"
 
@@ -29,4 +28,3 @@ type HttpResult struct {
 	Err         error     `json:"error"`
 	Status      int       `json:"httpStatus"`
 }
-
