+++ test/router/dnssec/dnssec_suite_test.go
@@ -8,9 +8,9 @@ package dnssec_test
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -19,16 +19,15 @@ package dnssec_test
  * under the License.
  */
 
-
 import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	"testing"
+	"flag"
 	"github.com/apache/trafficcontrol/test/router/dnssec"
 	"github.com/miekg/dns"
-	"flag"
 	"log"
+	"testing"
 )
 
 var d *dnssec.DnssecClient
@@ -36,8 +35,8 @@ var nameserver string
 var deliveryService string
 
 func init() {
-	flag.StringVar(&nameserver,"ns","changeit","ns is used to direct dns queries to a traffic router")
-	flag.StringVar(&deliveryService,"ds","changeit","ds is used to target some dns DS and DNS queries made by traffic router")
+	flag.StringVar(&nameserver, "ns", "changeit", "ns is used to direct dns queries to a traffic router")
+	flag.StringVar(&deliveryService, "ds", "changeit", "ds is used to target some dns DS and DNS queries made by traffic router")
 }
 
 var _ = BeforeSuite(func() {
@@ -46,7 +45,7 @@ var _ = BeforeSuite(func() {
 
 	Expect(nameserver).ToNot(Equal("changeit"), "Pass in a ns flag with the hostname of the traffic router")
 	Expect(deliveryService).ToNot(Equal("changeit"), "Pass in a ds flag with the dns label for a DNS delivery service")
-	log.Println("Nameserver",nameserver)
+	log.Println("Nameserver", nameserver)
 	log.Println("DeliveryService", deliveryService)
 })
