+++ test/router/dnssec/dnssec_test.go
@@ -8,9 +8,9 @@ package dnssec_test
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -18,7 +18,6 @@ package dnssec_test
  * specific language governing permissions and limitations
  * under the License.
  */
-
 
 import (
 	"github.com/apache/trafficcontrol/test/router/dnssec"
@@ -107,7 +106,6 @@ var _ = Describe("Dnssec", func() {
 
 		It("Uses KSK public key to verify ZSK RRSig", func() {
 			signedKeys := d.SigningData(nameserver, deliveryService)
-
 
 			count := 0
 			for _, signedZsk := range signedKeys.SignedZsks {
